Change syntax highlighting to use a state based approach
After finding another edge case and trying to update the old code to handle it, I decided to rewrite
it what I think is much more understandable manner, as a state machine. This makes it easier to
follow the flow of the code and debug issues. It should also make it more robust going forward.