backends: Adjust the backends for lazy-load vterm
Fixes: #135

The detection of whether vterm is present during init or not, before
making the functionality to load vterm is somewhat of a chicken-egg. This
change allows the vterm backend to be referenced during init. However,
loading of the vterm is still on demand based on the users
configuration.  Full details are in the GH issue.

Signed-off-by: Brian McGillion <bmg.avoin@gmail.com>